requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.4
django-extensions==1.8.1
django-test-without-migrations==0.6
python-decouple==3.0
pytz==2017.2
six==1.10.0
static3==0.7.0
django-debug-toolbar==1.8
gunicorn==19.4.3
psycopg2==2.7.3
sqlformatter
djangorestframework==3.6.4
markdown==2.6.9
django-filter==1.0.4

#django_select2==5.11.1
#django-selectable==1.0.0